requirements.txt
python-dotenv
pandas
PyPDF2
reportlab
tiktoken
chromadb
langchain
langchainhub
langchain-pinecone
langchain_community
langchain-openai
streamlit-lottie